requirements.txt
discord.py
fastapi
requests